Split apply code into multiple files
